Solution to Calculate Total and Average Marks of Students

+++ Calculate Total and Average Marks of Students@@67877d61dc107486ad5c2746/code.c
@@ -1,4 +1,3 @@
-// Your code here...
 #include <stdio.h>
 
 struct Student {
@@ -14,14 +13,15 @@ int main() {
     scanf("%d", &n);
 
     struct Student students[n];
-    int total_marks=0.0;
+    float total_marks = 0.0;  // Initialize total_marks as float
+
     // Input student details
     for (int i = 0; i < n; i++) {
         scanf("%d %s %f", &students[i].roll_number, students[i].name, &students[i].marks);
         total_marks += students[i].marks;
     }
 
-    // Calculate and print the average marks rounded to two decimal places
+    // Calculate and print the total and average marks rounded to two decimal places
     printf("Total Marks: %.2f\n", total_marks);
     printf("Average Marks: %.2f\n", total_marks / n);
